requirements.txt
requests==2.27.1
requests-toolbelt==0.9.1
streamlit==1.15.1
streamlit-aggrid==0.3.2
pandas==1.3.5
numpy==1.20.2
plotly==5.6.0
fredapi==0.5.0
PyYAML==5.4.1
xgboost==1.6.2